requirements.txt
click
PyYAML
requests
sdmx1
